Refactored /api/submissions
