--- Server/api/index/routes/index.js
@@ -1,13 +1,0 @@
-module.exports = {
-  method: 'GET',
-  path: '/',
-  config: {
-    auth: false
-  },
-  handler: (req, res) => {
-    res({
-      success: true,
-      message: 'Server is running!'
-    });
-  }
-};
