--- Server/api/submissions/handler/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  getSubmissions: require('./getSubmissions'),
-  addSubmission: require('./addSubmission'),
-  deleteSubmission: require('./id/deleteSubmission'),
-  updateSubmission: require('./id/updateSubmission')
-};
