--- Server/api/survey/handler/getSurvey.js
@@ -1,15 +1,0 @@
-'use strict';
-
-const Boom = require('boom');
-const Survey = require('../models/Survey');
-const User = require('../../users/models/User');
-
-module.exports = (req, res) => {
-  Survey.findOne({
-    user_id: req.Token.id
-  }, (err, survey) => {
-    res(((survey) ? survey : {
-      error: 'No Survey found!'
-    })).code(200);
-  });
-};
