--- Server/api/survey/handler/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  getSurvey: require('./getSurvey'),
-  addSurvey: require('./addSurvey'),
-  updateSurvey: require('./updateSurvey')
-};
